+++ client/src/pages/studyDetails/StudyDetails.js
@@ -21,24 +21,26 @@ import StudyNotebookTab from './StudyNotebookTab';
 import StudyConclusionsTab from "./StudyConclusionsTab";
 import StudyCommentsTab from "./StudyCommentsTab";
 import StudyTimelineTab from "./StudyTimelineTab";
-import AddToStudyCollectionModal
-  from "../../common/modals/AddToStudyCollectionModal";
+import AddToStudyCollectionModal from "../../common/modals/AddToStudyCollectionModal";
 import StudyCollectionsTab from "./StudyCollectionsTab";
 import PropTypes from "prop-types";
 import {useLocation, useNavigate} from "react-router-dom";
 import StudyDetailHeader from "./StudyDetailsHeader";
 import StudyOverviewTab from "./StudyOverviewTab";
 import StudyFileManagerTab from "./StudyFileManagerTab";
+import {useDispatch, useSelector} from "react-redux";
+import {setTab} from "../../redux/tabSlice";
 
 const StudyDetails = ({study, user, features}) => {
 
   const location = useLocation();
   const navigate = useNavigate();
-  const [selectedTab, setSelectedTab] = useState(location.hash.replace("#", "") || "overview");
   const [showCollectionModal, setShowCollectionModal] = useState(false);
+  const tab = useSelector(s => s.tab.value) || location.hash.replace("#", "") || "overview";
+  const dispatch = useDispatch();
 
   const handleTabSelect = (key) => {
-    setSelectedTab(key);
+    dispatch(setTab(key));
     navigate("#" + key);
   }
 
@@ -68,7 +70,7 @@ const StudyDetails = ({study, user, features}) => {
             {/* Tabs */}
             <Tabs
                 variant={"pills"}
-                activeKey={selectedTab}
+                activeKey={tab}
                 onSelect={handleTabSelect}
             >